chore: format code & use service split words
