+++ src/main/kotlin/plus/maa/backend/config/external/SegmentInfo.kt
@@ -1,0 +1,5 @@
+package plus.maa.backend.config.external
+
+data class SegmentInfo(
+    var path: String = "classpath:arknights.txt",
+)
